+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * [#360](https://github.com/clojure-emacs/clojure-mode/issues/360): `clojure-align` now reindents after aligning, which also fixes an issue with nested alignings.
 
+### New features
+
+* [#362](https://github.com/clojure-emacs/clojure-mode/issues/362): New custom option `clojure-indent-style` offers 3 different ways to indent code.
+
 ## 5.1.0 (04/01/2015)
 
 ### New features